Extract LayoutFileManager class
